Codebase Refactor
