+++ DeepQuant/QuantManipulation/QuantDequantNodes.py
@@ -4,24 +4,12 @@
 #
 # Federico Brancasi <fbrancasi@ethz.ch>
 
-"""
-Basic implementation of Quant and Dequant modules.
-"""
-
 import torch
 import torch.nn as nn
-from typing import Any, Optional, Union
+from typing import Optional
 
 
 class Quant(nn.Module):
-    """
-    Fake-quant module that applies a "saturating" approach using scale, zero_point, bit_width,
-    and signedness parameters extracted from a Brevitas parameter dictionary.
-
-    This module simulates quantization effects on tensors by scaling, shifting, rounding,
-    and clamping their values.
-    """
-
     def __init__(
         self,
         original_module: nn.Module,
@@ -30,16 +18,6 @@ class Quant(nn.Module):
         bit_width: float,
         signed: Optional[bool] = True,
     ) -> None:
-        """
-        Initialize the Quant module.
-
-        Args:
-            original_module: The original Brevitas quant module (kept for reference).
-            scale: Scale factor used for quantization.
-            zero_point: Zero-point used for quantization.
-            bit_width: Bit width for the quantized representation (e.g., 8.0, 32.0).
-            signed: Boolean flag indicating if quantization is signed.
-        """
         super().__init__()
         self.original_module = original_module
         self.scale = scale
@@ -60,22 +38,6 @@ class Quant(nn.Module):
             self.max_val = None
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
-        """
-        Apply fake quantization to the input tensor.
-
-        The quantization process is as follows:
-          1) Scale the input tensor by 1/scale.
-          2) Shift the scaled tensor by the zero_point.
-          3) Round the shifted tensor to the nearest integer.
-          4) Clamp the rounded tensor to the representable range based on bit_width
-             and signedness.
-
-        Args:
-            x: Input tensor.
-
-        Returns:
-            The fake quantized tensor.
-        """
         if self.scale is None or self.zero_point is None:
             return x
 
@@ -88,10 +50,6 @@ class Quant(nn.Module):
 
 
 class Dequant(nn.Module):
-    """
-    Dequant module that re-applies scale and zero_point to invert the quantization effect.
-    """
-
     def __init__(
         self,
         original_module: nn.Module,
@@ -100,16 +58,6 @@ class Dequant(nn.Module):
         bit_width: float,
         signed: Optional[bool] = True,
     ) -> None:
-        """
-        Initialize the Dequant module.
-
-        Args:
-            original_module: The original Brevitas quant module.
-            scale: Scale factor from extracted parameters.
-            zero_point: Zero-point from extracted parameters.
-            bit_width: Bit width from extracted parameters.
-            signed: Boolean flag indicating if quantization is signed.
-        """
         super().__init__()
         self.original_module = original_module
         self.scale = scale
@@ -118,15 +66,6 @@ class Dequant(nn.Module):
         self.signed = signed
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
-        """
-        Undo the fake quantization by reversing the shift and scale.
-
-        Args:
-            x: Input tensor.
-
-        Returns:
-            The dequantized tensor.
-        """
         if self.scale is None or self.zero_point is None:
             return x
         x_dequant = (x - self.zero_point) * self.scale
