+++ DeepQuant/__init__.py
@@ -1,0 +1,9 @@
+# Copyright 2025 ETH Zurich and University of Bologna.
+# Licensed under the Apache License, Version 2.0, see LICENSE for details.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Federico Brancasi <fbrancasi@ethz.ch>
+
+from DeepQuant.Export import exportQuantModel
+
+__all__ = ["exportQuantModel"]
